HTTP Handler for functions
The function handler will be the handler for each http trigger.  It
will look up the function service map to find a service for the
function (and call the pool manager to create a new service if one
doesn't already exist, though this isn't implemented yet).  It'll then
proxy the request to the service.