Create a deployment to adopt the function pod
When specializing a pod, remove its ownerReference to the generic
pool, and create a deployment that will own it. Now all we have to do
to autoscale is to actually tweak the number of replicas on this
deployment.

Also, this change is missing cleanup for both the deployment and
service.