Add EnvironmentReference to Packages
Since packages have source and deployment, and can be shared by
functions, packages correspond to an environment too. Without this,
the environment of a package would be determined by what function
references it, which is not a great idea.

This also adds json labels to types that didn't have them.