ArchiveLiteralSizeLimit: Use Constant Instead Hard Code
This use constant to instead hard code for ArchiveLiteralSizeLimit. That ensure if the limitation
value is changed in cli, the controller server can follow it. Otherwise, the condition will return
error, and make error HTTP response.

Later, we should delete it in fission-cli, and make it configurable it chart, to work in fission-
servers only.

Signed-off-by: xiekeyang <keyang.xie@gmail.com>