Bug Fix: Fission Client should exit 1 when its environment not found
The older mechanism, in the case of environment not exist, allows program to go on, and create a bad function.
This let customer get a incorrect result, and set up a unworkable function.

Actually we should stop the program when environment not found, and return failed result, with the error message.

Signed-off-by: xiekeyang <keyang.xie@gmail.com>