Allow CLI to wait for router controllers
When a script or a human user creates some resources (functions,
triggers), the router notices these changes and updates its config.
However the user or script has no way of knowing when the router has
completed this task.

This change adds: (1) a router API endpoint that returns the highest
among the ResourceVersions of all functions and httptriggers it has
seen, and (2) A cli change that writes the RV of any created or
updated object into a file (3) A cli change that waits for the
router's reported RV to match the one written into the file.

The whole idea is a violation of abstraction layers.  It assumes that
RVs are monotonically increasing uint64s while the K8s API only
guarantees that they are unique strings (not even numbers).  However
K8s gets these numbers from the etcd index, which is (I think) the Raft
log index, which does increase monotonically (I think).