Replace several test sleeps with fission/kubectl wait
The fission CLI now lets us wait for the router to catch up, so we can
eliminate lots of "wait for router" sleeps.

Also, Kubernetes 1.11 introduces "kubectl wait" which lets us wait for
a deployment to be ready with a timeout.  We use this in a few places
that we were sleeping or polling for the environment pods to be up.