Fix executor tries to create same name deployment
The root cause of the issue was introduced by PR https://github.com/fission/fission/pull/1009/files .
To be short, even the CRD of environment was delete, it still takes time for executor (poolmgr) to destroy env pool.
So, when a user can create a same name env, following logs will be shown.
```
2019/01/08 11:01:26 Error creating deployment for nodejs-poolmgr-default in kubernetes, err: deployments.extensions "nodejs-poolmgr-default" already exists
2019/01/08 11:01:26 eager-create pool failed: deployments.extensions "nodejs-poolmgr-default" already exists
```
In our cases, the previous test creates an env and delete it when test finished, then the next one creates a same name env, but failed to create pool due to the deploy name conflict like log above. So the executor selects the pod from the first created env pool. Then, executor starts to delete the env pool, and makes the pod state became Termination state.

To fix this problem, a unique of deployment will be return after this PR to prevent the name conflict.