Remove prometheus server connectivity test during controller initialization
The prometheus server may not up in a short time or a cluster doesn’t have a storage class config. Because of this, the connectivity test will increase controller initialization time and so the controller will fail at readiness probe detection. Since we don’t want to delay the time that controller comes up also canary config manager prints error log when it failed at connecting to prometheus server, we remove the connectivity test here.