Issue #1258: Allow empty repository tag in fission chart values
This fix allows the empty repository tag in values.yaml file.
The changes are done in the templates of deployment and fluentbit where repository tag will be added
in the docker image name only when specified with some value in case of empty repository tag image name
contain the name of image and tag only.