Allow using URL as archive source when creating functions
In case of large files, it takes a long time for user to download
the source from the URL and then upload it to StorgeSvc through
CLI.

This PR allows a user to use URL as the function source when creating a function
and provides a new flag "--keeparchiveurl" to let the user to decided
whether the CLI should download the file first or store the file URL in the
archive directly. If "--keeparchiveurl" is true, then no checksum will be
generated, it's the user's responsibility to ensure the file won't be changed.