Fix helm shows "Not a table" issue when install Fission
Helm treats the single line "extraCoreComponentPodConfig" in
values.yaml as a key value pair config.

```
extraCoreComponentPodConfig:
```

And so when a user tries to install helm with providing config
like following, it failed.

```
extraCoreComponentPodConfig:
  nodeSelector:
    serverless: fission
```

The root cause is that helm treats config in default values.yaml
and providing config as two different types of config, that is
key-value vs. value table. And so it's unable to merge these two
configs.

This PR comments out the single line config in default values.yaml
to prevent the issue.