Fix logger init container has no permission to create directory
On machine that doesn't allow container to create sub-directory
in the directory that mounts into container, the logger enter CrashLoopBackOff.

This PR change the location of symlink path to prevent permission
problem when init container trying to create directory and a new PV
is added for fluent-bit to store it's database file.